CODE RUB: Move LabOrchestrationService and Tests Under Named Folders
