--- DMX.Core.Api.Tests.Unit/Services/Orchestrations/LabCommands/Theory.cs
@@ -1,10 +1,0 @@
-﻿// ---------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// ---------------------------------------------------------------
-
-namespace DMX.Core.Api.Tests.Unit.Services.Orchestrations.LabCommands
-{
-    internal class Theory<T>
-    {
-    }
-}
