+++ DMX.Core.Api.Tests.Unit/Services/Orchestrations/Labs/LabOrchestrationServiceTests.Validations.Add.cs
@@ -9,7 +9,7 @@ using FluentAssertions;
 using Moq;
 using Xunit;
 
-namespace DMX.Core.Api.Tests.Unit.Services.Orchestrations
+namespace DMX.Core.Api.Tests.Unit.Services.Orchestrations.Labs
 {
     public partial class LabOrchestrationServiceTests
     {
