+++ DMX.Core.Api.Tests.Unit/Services/Orchestrations/Labs/LabOrchestrationServiceTests.Validations.RetrieveById.cs
@@ -10,44 +10,43 @@ using FluentAssertions;
 using Moq;
 using Xunit;
 
-namespace DMX.Core.Api.Tests.Unit.Services.Orchestrations
+namespace DMX.Core.Api.Tests.Unit.Services.Orchestrations.Labs
 {
     public partial class LabOrchestrationServiceTests
     {
         [Fact]
-        public async Task ShouldThrowValidationExceptionOnRemoveIfLabIdIsInvalidAndLogItAsync()
+        public async Task ShouldThrowValidationExceptionOnRetrieveByIdIfLabIdIsInvalidAndLogItAsync()
         {
             // given
             var invalidLabId = Guid.Empty;
-            var invalidLabOrchestrationException = new InvalidLabOrchestrationException();
+            var invalidLabException = new InvalidLabOrchestrationException();
 
-            invalidLabOrchestrationException.AddData(
+            invalidLabException.AddData(
                 key: nameof(Lab.Id),
                 values: "Id is required");
 
             var expectedLabOrchestrationValidationException =
-                new LabOrchestrationValidationException(invalidLabOrchestrationException);
+                new LabOrchestrationValidationException(invalidLabException);
 
             // when
             ValueTask<Lab> actualLabTask =
-                this.labOrchestrationService.RemoveLabByIdAsync(invalidLabId);
+                this.labOrchestrationService.RetrieveLabByIdAsync(invalidLabId);
 
-            LabOrchestrationValidationException
-                actualLabOrchestrationValidationException =
-                    await Assert.ThrowsAsync<LabOrchestrationValidationException>(
-                        actualLabTask.AsTask);
+            LabOrchestrationValidationException actualLabOrchestrationValidationException =
+                await Assert.ThrowsAsync<LabOrchestrationValidationException>(
+                    actualLabTask.AsTask);
 
             // then
             actualLabOrchestrationValidationException.Should()
                 .BeEquivalentTo(expectedLabOrchestrationValidationException);
 
             this.loggingBrokerMock.Verify(broker =>
-                broker.LogError(It.Is(SameExceptionAs(
-                    expectedLabOrchestrationValidationException))),
+                 broker.LogError(It.Is(SameExceptionAs(
+                     expectedLabOrchestrationValidationException))),
                         Times.Once);
 
             this.labServiceMock.Verify(service =>
-                service.RemoveLabByIdAsync(It.IsAny<Guid>()),
+                service.RetrieveLabByIdAsync(It.IsAny<Guid>()),
                     Times.Never);
 
             this.loggingBrokerMock.VerifyNoOtherCalls();
