+++ DMX.Core.Api/Controllers/LabsController.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using DMX.Core.Api.Models.Foundations.Labs;
 using DMX.Core.Api.Models.Foundations.Labs.Exceptions;
 using DMX.Core.Api.Models.Orchestrations.Labs.Exceptions;
-using DMX.Core.Api.Services.Orchestrations;
+using DMX.Core.Api.Services.Orchestrations.Labs;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using RESTFulSense.Controllers;
